requirements.txt
alembic==1.9.4
aniso8601==9.0.1
anyio==3.6.2
bcrypt==3.2.0
blinker==1.4
cachetools==4.2.0
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
charset-normalizer==2.1.1
click==7.1.2
cmake==3.26.0
contourpy==1.0.7
cycler==0.11.0
dnspython==2.0.0
email-validator==1.1.2
filelock==3.10.2
Flask==1.1.2
Flask-Bcrypt==0.7.1
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==4.0.4
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
fonttools==4.39.3
google-api-core==1.24.1
google-api-python-client==1.12.8
google-auth==1.24.0
google-auth-httplib2==0.0.4
google-auth-oauthlib==0.4.2
googleapis-common-protos==1.52.0
h11==0.14.0
httpcore==0.16.3
httplib2==0.18.1
httpx==0.23.3
huggingface-hub==0.13.3
idna==2.10
importlib-metadata==6.0.0
importlib-resources==5.12.0
itsdangerous==1.1.0
Jinja2==2.11.2
jsonify==0.5
kiwisolver==1.4.4
lit==16.0.0
Mako==1.2.4
MarkupSafe==1.1.1
matplotlib==3.7.1
mpmath==1.3.0
networkx==3.0
numpy==1.23.2
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauth2client==4.1.3
oauthlib==3.1.0
opencv-python==4.7.0.72
packaging==23.0
Pillow==8.0.1
protobuf==3.14.0
psycopg2-binary==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pydicom==2.3.0
pyorthanc==1.11.5
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2020.5
PyYAML==6.0
requests==2.28.1
requests-oauthlib==1.3.0
rfc3986==1.5.0
rsa==4.6
six==1.15.0
sniffio==1.3.0
SQLAlchemy==1.3.20
SQLAlchemy-Utils==0.38.3
sympy==1.11.1
timm==0.6.12
torch==2.0.0
torchvision==0.15.1
tqdm==4.65.0
triton==2.0.0
typing-extensions==4.5.0
uritemplate==3.0.1
urllib3==1.26.2
Werkzeug==1.0.1
wtf==0.1
WTForms==2.3.3
yacs==0.1.8
zipp==3.15.0
